Rename package from math to nmath
