+++ utils/nmath/math.go
@@ -1,4 +1,4 @@
-package math
+package nmath
 
 import "math"
